Combine Interface: TSS; Actually almost completely arch-specific
